refactor: move outbound-proxy test files to assistant/src/outbound-proxy/__tests__/
Move 13 test files that test outbound-proxy internals from scattered locations
(assistant/src/__tests__/, assistant/src/tools/network/script-proxy/__tests__/)
into assistant/src/outbound-proxy/__tests__/ so they live next to the code they
test. Import paths updated to use ../index.js for the outbound-proxy barrel and
../../ for assistant module references.

Also updates the migration tracking doc with completed status for categories 3-4.
