+++ assistant/src/outbound-proxy/__tests__/fixtures/credential-security-fixtures.ts
