+++ assistant/src/outbound-proxy/__tests__/script-proxy-certs.test.ts
@@ -7,7 +7,7 @@ import {
   ensureLocalCA,
   getCAPath,
   issueLeafCert,
-} from "../outbound-proxy/index.js";
+} from "../index.js";
 
 let dataDir: string;
